Authorization

+++ app/Http/Controllers/TagController.php
@@ -4,9 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TagController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['index']); //middleware bisa juga di web.php, di sini bisa custom yang except
+        $this->middleware('admin')->except(['index']);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -41,7 +47,7 @@ class TagController extends Controller
             'name' => 'required',
             'style' => 'required',
         ]);
-        
+
         $tag = new Tag([
             'name' => $request['name'],
             'style' => $request['style']
@@ -71,6 +77,8 @@ class TagController extends Controller
      */
     public function edit(Tag $tag)
     {
+        abort_unless(Gate::allows('udpate', $tag), 403);
+
         return view('tag.edit')->with([
             'tag' => $tag
         ]);
@@ -85,11 +93,13 @@ class TagController extends Controller
      */
     public function update(Request $request, Tag $tag)
     {
+        abort_unless(Gate::allows('udpate', $tag), 403);
+
         $request->validate([
             'name' => 'required',
             'style' => 'required',
         ]);
-        
+
         $tag->update([
             'name' => $request['name'],
             'style' => $request['style']
@@ -107,6 +117,8 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        abort_unless(Gate::allows('delete', $tag), 403);
+
         $oldName = $tag->name;
         $tag->delete();
         return $this->index()->with([
